multi: pull reputation tracking out into separate file
